added functionality to nav links and search bar. Had to put other components into content component to make the funtionality work
