+++ src/App.js
@@ -1,11 +1,9 @@
-import Nav from "./components/Nav.js";
 import Content from "./components/Content.js";
 import Footer from "./components/Footer.js";
 
 function App() {
   return (
     <div>
-      <Nav />
       <Content />
       <Footer />
     </div>
